Chunk index buffers use GL_UNSIGNED_INT instead of GL_UNSIGNED_SHORT to allow experimentation with larger chunk sizes and to avoid crashes when chunks contain a lot of geometry.

+++ GutsyStorm/GSChunkGeometryData.h
@@ -26,7 +26,7 @@
     GLfloat *normsBuffer;
     GLfloat *texCoordsBuffer;
     GLfloat *colorBuffer;
-    GLushort *indexBuffer;
+    GLuint *indexBuffer;
     
 @public
     GSVector3 corners[8];
